Generate byte sizes as 64-bit, so a quota above 2 GiB can be declared
LinkML's gen-java maps integer to int/Integer and has no xsd:long mapping, so every byte-size
attribute was 32-bit and a 10 GiB PersonalSpace quota overflowed at load, refusing the whole
corpus. A ByteCount type (typeof integer, uri xsd:long) now ranges the nine byte-size attributes,
and scripts/generate/gen_java.py adds the xsd:long -> Long/long mapping gen-java lacks. JSON
Schema is unchanged (integer). metamodel/VERSION and jumo-model move to 0.64.0.

Jumo-Work-Order: metamodel-byte-sizes-overflow-java-int
Jumo-Role: implementer
Jumo-Contract-Revision: bb469c63b861c1ce536520838bf6e13a5d397b71

+++ modules/jumo-model/src/main/java/dev/jumo/model/MachineHostDefinitionSpec.java
@@ -7,7 +7,7 @@ import java.util.Objects;
 /**
  * Specification for a MachineHostDefinition contract.
  */
-public record MachineHostDefinitionSpec(String ownerRealm, String driver, String domainPrefix, String storagePool, String bridgeNetwork, Integer memoryBytes, Integer vcpuCount)  {
+public record MachineHostDefinitionSpec(String ownerRealm, String driver, String domainPrefix, String storagePool, String bridgeNetwork, Long memoryBytes, Integer vcpuCount)  {
 
     public static Builder builder() {
         return new Builder();
@@ -20,7 +20,7 @@ public record MachineHostDefinitionSpec(String ownerRealm, String driver, String
         private String domainPrefix = "";
         private String storagePool = "";
         private String bridgeNetwork = "";
-        private Integer memoryBytes = null;
+        private Long memoryBytes = null;
         private Integer vcpuCount = null;
 
 
@@ -49,7 +49,7 @@ public record MachineHostDefinitionSpec(String ownerRealm, String driver, String
             return this;
         }
 
-        public Builder memoryBytes(Integer memoryBytes) {
+        public Builder memoryBytes(Long memoryBytes) {
             this.memoryBytes = Objects.requireNonNull(memoryBytes);
             return this;
         }
